backend/requirements.txt
Django==2.2
gunicorn==19.3.0
redis==2.10.6
djangorestframework
django-cors-headers
django-extra-fields
Pillow
django-filter
djangorestframework_simplejwt
PyMySQL
mysqlclient
channels==1.1.8
asgi-redis
daphne==1.3
pyjwt==1.6.4
sendgrid
xmltodict